[cmd] Move legacy uninstall logic
